Move `lib` code away from `model`
