+++ src/lib/pages/podcast-gallery/index.ts
@@ -1,2 +1,2 @@
 export { default as PodcastGalleryPage } from './ui/index.svelte';
-export { loadPodcasts } from './model/load-podcasts';
+export { loadPodcasts } from './lib/load-podcasts';
